cleaning code + some new implementations

+++ parser/email_parser.py
@@ -3,12 +3,19 @@ import os
 from email import policy
 from email.parser import BytesParser
 import mailbox
+from datetime import datetime
+from typing import Optional, Tuple, List, Dict, Union, Any
+
+from config.email_parser_constants import EmailParserConstants
+from config.email_constants import *
+from utils.file_text_extractor import FileTextExtractor
+
 from .iemail_parser import IEmailParser
 from utils.string_cleaner import StringCleaner
 from utils.date_transformer import DateTransformer
-from hasher.hasher import Hasher
 from utils.logging_setup import log_email_parser
-from config.email_parser_constants import EmailParserConstants
+from hasher.hasher import Hasher
+
 
 class EmailParser(IEmailParser):
     def __init__(self, attachments_directory=None):
@@ -20,7 +27,7 @@ class EmailParser(IEmailParser):
         else:
             self.attachments_directory = EmailParserConstants.ATTACHMENTS_DIRECTORY
 
-    def parse_email(self, email_content):
+    def parse_email(self, email_content: bytes) -> dict:
         """
         Analyse le contenu d'un email et retourne un dictionnaire avec les données pertinentes.
 
@@ -32,59 +39,37 @@ class EmailParser(IEmailParser):
         """
         log_email_parser.info("Func: parse_email")
         msg = BytesParser(policy=policy.default).parsebytes(email_content)
-        # print(f"msg keys: {msg.keys()}")
         email_id = self.hasher.hash_string(data=msg.as_bytes())
         log_email_parser.debug(f"Func: parse_email for email_id: {email_id}")
-        # print(f"Date: {msg['date']}")
         date = self._transform_date(msg['date'])
         log_email_parser.debug(f"Func: parse_email with date: {date}")
         body, attachments = self.extract_body_and_attachments(msg=msg)
-
 
         from_name, from_address = self._parse_names_addresses(data=msg['from'])
         to_names, to_addresses = self._parse_names_addresses(data=msg['to'])
         cc_names, cc_addresses = self._parse_names_addresses(data=msg['cc'])
         bcc_names, bcc_addresses = self._parse_names_addresses(data=msg['bcc'])
 
-
         return {
-            'email_id': email_id,
-            'from_name': from_name,
-            'from_address': from_address,
-            'subject': msg['subject'],
-            'date_str': date.strftime('%Y-%m-%d %H:%M:%S') if date else None,
-            'date_obj': date if date else None,
-            'date_iso': date.isoformat() if date else None,
-            'timestamp': date.timestamp() if date else None,
-            'to_names': to_names,
-            'to_addresses': to_addresses,
-            'cc_names': cc_names,
-            'cc_addresses': cc_addresses,
-            'bcc_names': bcc_names,
-            'bcc_addresses': bcc_addresses,
-            'body': body,
-            'attachments': attachments
+            EMAIL_ID: email_id,
+            FROM_NAME: from_name,
+            FROM_ADDRESS: from_address,
+            SUBJECT: msg['subject'],
+            DATE_STR: date.strftime(DATETIME_FORMAT) if date else None,
+            DATE_OBJ: date if date else None,
+            DATE_ISO: date.isoformat() if date else None,
+            TIMESTAMP: date.timestamp() if date else None,
+            TO_NAMES: to_names,
+            TO_ADDRESSES: to_addresses,
+            CC_NAMES: cc_names,
+            CC_ADDRESSES: cc_addresses,
+            BCC_NAMES: bcc_names,
+            BCC_ADDRESSES: bcc_addresses,
+            BODY: body,
+            ATTACHMENTS: attachments
         }
 
-    def parse_mbox(self, file_path):
-        """
-        Analyse le contenu d'un fichier mbox et retourne une liste de dictionnaires avec les données des emails.
-
-        Paramètres:
-        file_path (str): Le chemin du fichier mbox.
-
-        Retourne:
-        list: Une liste de dictionnaires contenant les données des emails.
-        """
-        log_email_parser.info("Func: parse_mbox")
-        mbox = mailbox.mbox(file_path)
-        emails = []
-        for message in mbox:
-            email = self.parse_email(message.as_bytes())
-            emails.append(email)
-        return emails
-
-    def extract_body_and_attachments(self, msg):
+    def extract_body_and_attachments(self, msg: BytesParser) -> Tuple[Optional[str], List[Dict[str, Union[str, bytes]]]]:
         log_email_parser.info("Func: extract_body_and_attachments")
         body = None
         attachments = []
@@ -120,15 +105,14 @@ class EmailParser(IEmailParser):
                 if content is not None:
                     attachment_id = self.hasher.hash_string(data=content)
                     log_email_parser.debug(f"Func: extract_body_and_attachments, attachment_id: {attachment_id}")
-                    self._download_attachment(content=content, attachment_id=attachment_id, filename=filename)
-                    # ToDo !!!
-                    #extractor = AttachmentTextExtractor(file_content=content)
-                    #extracted_text = extractor.extract_text()
+                    filepath = self._download_attachment(content=content, attachment_id=attachment_id, filename=filename)
+                    file_text_extractor = FileTextExtractor(file_path=filepath)
+                    extracted_text = file_text_extractor.extract_text()
                     attachments.append({
                         'attachment_id': attachment_id,
                         'filename': filename,
                         'content': content,
-                        'extracted_text': "extracted_text" # Todo !!!!
+                        'extracted_text': extracted_text  # Todo: tests à faire !
                     })
                     del content
                     gc.collect()
@@ -137,33 +121,30 @@ class EmailParser(IEmailParser):
                     print(f"Warning: Attachment {filename} has no content and was skipped.")
         return body, attachments
 
-    def _download_attachment(self, content, attachment_id, filename, save_directory="attachments"):
-        log_email_parser.info("Func: _download_attachment")
+    def _download_attachment(self, content: Any, attachment_id: str, filename: str) -> None:
         try:
-            if not os.path.exists(save_directory):
-                os.makedirs(save_directory)
-
+            self.sc.create_directory(dir_path=self.attachments_directory)
             new_filename = self.sc.rename_file(filename=filename, new_name=attachment_id)
-            filepath = os.path.join(save_directory, new_filename)
-
-            with open(filepath, 'wb') as f:
-                f.write(content)
-
-            #print(f"Attachment saved to {filepath}")
+            filepath = os.path.join(self.attachments_directory, new_filename)
+            if not self.sc.file_exists(filepath):
+                log_email_parser.info(f"Download attachment {filepath}")
+                with open(filepath, 'wb') as f:
+                    f.write(content)
+            else:
+                log_email_parser.debug(f"Attachment {filename} already exists")
+            return filepath
         except Exception as e:
             raise Exception(f"{filename}: {e}")
 
-
-    def _parse_names_addresses(self, data):
-        #log_email_parser.info("Func: _parse_names_addresses")
+    def _parse_names_addresses(self, data: str) -> Tuple[List[str], List[str]]:
+        # log_email_parser.info("Func: _parse_names_addresses")
         list_names_and_addresses = self.split_name_address(fieldvalue=data)
         names, addresses = self.separate_names_and_addresses_from_list(list_names_and_addresses)
         names = self.sc.replace_chars_by_char(data=names, current_chars={'.'}, new_char=' ')
         return names, addresses
 
-
-    def _transform_date(self, date_input):
-        #log_email_parser.info("Func: _transform_date")
+    def _transform_date(self, date_input: str) -> datetime:
+        # log_email_parser.info("Func: _transform_date")
         date_obj = self.dt.parse_email_date(date_input=date_input)
         date_obj = self.dt.change_time_shift(date_input=date_obj)
         return date_obj
@@ -176,7 +157,7 @@ class EmailParser(IEmailParser):
         Sometimes a name is the e-mail address, so the person field will contain the address twice,
         and therefore twice '@'.
         """
-        #log_email_parser.info("Func: split_name_address")
+        # log_email_parser.info("Func: split_name_address")
         if fieldvalue is None:
             return None
         if not isinstance(fieldvalue, str):
@@ -191,11 +172,8 @@ class EmailParser(IEmailParser):
                 field = ''
         if split_field == []:
             split_field.append(fieldvalue)
-
         names_address = []
-        # split_field = fieldvalue.split(',')
         for field in split_field:
-            # print(f"Address field: {field}")
             last_space_index = field.rfind(' ')
 
             if last_space_index == -1:
@@ -203,7 +181,7 @@ class EmailParser(IEmailParser):
                 names_address.append(('', address))
             else:
                 name = field[:last_space_index]
-                address = field[last_space_index+1:]
+                address = field[last_space_index + 1:]
                 name = self.sc.remove_chars(name)
                 address = self.sc.remove_chars(address)
                 if name == address:
@@ -212,11 +190,11 @@ class EmailParser(IEmailParser):
                     names_address.append((name, address))
         return names_address
 
-
-    def separate_names_and_addresses_from_list(self, list_of_name_address_tuple: list) -> tuple:
+    def separate_names_and_addresses_from_list(self, list_of_name_address_tuple: List[Tuple[str, str]]) -> Tuple[List[str], List[str]]:
         """For e-mail address fields"""
-        #log_email_parser.info("Func: separate_names_and_addresses_from_list")
+        # log_email_parser.info("Func: separate_names_and_addresses_from_list")
         if not list_of_name_address_tuple:
             return [], []
         names, addresses = zip(*list_of_name_address_tuple)
         return list(names), list(addresses)
+
